move sections into folder
