+++ src/renderer/components/SettingsSection/Customization.tsx
