+++ src/renderer/components/SettingsSection/index.tsx
@@ -3,8 +3,8 @@ import GeneralSettings from 'renderer/components/SettingsSection/General';
 import { Redirect, Route } from "react-router-dom";
 import { AboutSettings } from "renderer/components/SettingsSection/About";
 import { SideBar, SideBarLink, SideBarTitle } from "renderer/components/SideBar";
-import CustomizationSettings from '../CustomizationSettings';
-import DownloadSettings from '../DownloadSettings';
+import CustomizationSettings from './Customization';
+import DownloadSettings from './Download';
 
 export const SettingsSection = (): JSX.Element => {
     return (
